:sparkles: feat: tratativa de erro para delete de objetos com associações

+++ src/main/java/codifica/eleve/core/application/usecase/pet/DeletePetUseCase.java
@@ -1,14 +1,17 @@
 package codifica.eleve.core.application.usecase.pet;
 
+import codifica.eleve.core.domain.agenda.AgendaRepository;
 import codifica.eleve.core.domain.pet.PetRepository;
+import codifica.eleve.core.domain.shared.exceptions.ConflictException;
 import codifica.eleve.core.domain.shared.exceptions.NotFoundException;
 
 public class DeletePetUseCase {
     private final PetRepository petRepository;
-    // private final AgendaRepository agendaRepository;
+    private final AgendaRepository agendaRepository;
 
-    public DeletePetUseCase(PetRepository petRepository) {
+    public DeletePetUseCase(PetRepository petRepository, AgendaRepository agendaRepository) {
         this.petRepository = petRepository;
+        this.agendaRepository = agendaRepository;
     }
 
     public void execute(Integer id) {
@@ -16,9 +19,9 @@ public class DeletePetUseCase {
             throw new NotFoundException("Pet não encontrado.");
         }
 
-        // if (agendaRepository.existsByPetId(id)) {
-        //     throw new ConflictException("Não é possível deletar pets que possuem agendas associadas.");
-        // }
+        if (agendaRepository.existsByPetId(id)) {
+            throw new ConflictException("Não é possível deletar pets que possuem agendas associadas.");
+        }
 
         petRepository.deleteById(id);
     }
